enhance marker colors
