+++ src/pages/components/map-related/MapLayers/MapIndexLayer.tsx
@@ -5,8 +5,8 @@ import type { ReactNode } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router'
 import type { PositionGroup } from '../map-types'
-import { plannedRouteLineColor, plannedRouteStopMarkerPath } from '../MapContent'
 import { MapIndex } from '../MapIndex'
+import { plannedRouteLineColor, plannedRouteStopMarkerPath } from '../mapMarkers'
 import { MapSpeedIndex } from '../MapSpeedIndex'
 import { SPEED_BANDS, VehicleBearingGlyph } from '../vehicleBearingGlyph'
