+++ src/pages/components/map-related/map-types.ts
@@ -1,7 +1,9 @@
 import type { SiriVehicleLocationWithRelatedPydanticModel } from '@hasadna/open-bus-api-client'
 import { BusStop } from 'src/model/busStop'
 
-export const ROUTE_COLORS = ['#f97316', '#3b82f6', '#22c55e', '#a855f7', '#ef4444']
+/** One per concurrent ride on the map, cycled. Deliberately no red: red is the ping speed ramp's
+ * (`.ping-arrow--slow`), and red arrows over a red line read as neither. */
+export const ROUTE_COLORS = ['#f97316', '#3b82f6', '#22c55e', '#a855f7', '#0891b2']
 
 export interface Point {
   loc: [number, number]
