+++ src/pages/components/map-related/vehicleBearingGlyph.test.ts
@@ -3,6 +3,7 @@ import {
   bearingZIndex,
   BOOKEND_Z_INDEX,
   isStanding,
+  rideEndSvgMarkup,
   SPEED_BAND_MAX,
   SPEED_BANDS,
   speedBand,
@@ -65,14 +66,22 @@ describe('arrowSvgMarkup', () => {
   it('paints nothing itself, leaving the colours to the stylesheet that can theme them', () => {
     const markup = arrowSvgMarkup(45, 0)
 
-    expect(markup).toContain('class="ping-arrow"')
+    expect(markup).toContain('ping-arrow')
     expect(markup).not.toContain('fill=')
     expect(markup).not.toContain('stroke=')
   })
 
-  it('inks the slow half of the bands solid and hollows the fast half out', () => {
-    expect(arrowSvgMarkup(0, 0)).not.toContain('ping-arrow--outline')
-    expect(arrowSvgMarkup(0, TOP)).toContain('ping-arrow--outline')
+  it('reddens the slow half and only the slow half', () => {
+    expect(arrowSvgMarkup(0, 0)).toContain('ping-arrow--slow')
+    expect(arrowSvgMarkup(0, TOP)).not.toContain('ping-arrow--slow')
+  })
+
+  // The regression this ramp's floor exists to prevent: the slowest band is both the smallest
+  // arrow and the one worth finding, and below ~0.6 it is a speck between its neighbours.
+  it('keeps the slowest arrow big enough to find, not merely big enough to draw', () => {
+    const scale = Number(/scale\(([\d.]+)\)/.exec(arrowSvgMarkup(0, 0))![1])
+
+    expect(scale).toBeGreaterThanOrEqual(0.6)
   })
 })
 
@@ -85,6 +94,12 @@ describe('standingSvgMarkup', () => {
   it('drops the needle when there is no bearing, rather than inventing a heading', () => {
     expect(standingSvgMarkup(undefined)).not.toContain('rotate(')
     expect(standingSvgMarkup(undefined)).toContain('ping-badge')
+  })
+
+  // A parked bus is a speed reading and joins the red family; the end of a ride is not.
+  it('rims the parked bus in the slow ramp colour, which the ride-end badge must not take', () => {
+    expect(standingSvgMarkup(0)).toContain('ping-badge--standing')
+    expect(rideEndSvgMarkup()).not.toContain('ping-badge--standing')
   })
 })
